backend/pyproject.toml
[tool.poetry]
name = "wanllmdb-backend"
version = "0.1.0"
description = "wanLLMDB Backend API"
authors = ["wanLLMDB Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "~3.2.2"
python-multipart = "^0.0.6"
redis = "^5.0.1"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
slowapi = "^0.1.9"
psutil = "^5.9.6"
email-validator = "^2.3.0"
minio = "^7.2.18"

[tool.poetry.group.dev.dependencies]
optuna = "^3.6.0"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^23.12.1"
ruff = "^0.1.11"
mypy = "^1.8.0"
pre-commit = "^3.6.0"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C4"]
ignore = []
fixable = ["ALL"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
